your commit message

+++ kernels/count_alignment.cuh
@@ -4,8 +4,8 @@
 #include "cuda_runtime.h"
 #include "device_launch_parameters.h"
 
-void count_alignment_cuda(int batch_size, uint64_t local_nnz_count, uint64_t* mattuples0, uint64_t* mattuples1, uint64_t* cks_count,
-    uint64_t col_offset,uint64_t row_offset, int ckthr,uint64_t* align_batch, uint64_t* elimi_batch);
+void count_alignment_cuda(int batch_size, int local_nnz_count, int* mattuples0, int* mattuples1, int* cks_count,
+    int col_offset,int row_offset, int ckthr, int* align_batch, int* elimi_batch);
 
 void test();
